+++ test/process_manager_test.dart
@@ -64,9 +64,25 @@ void main() {
       );
       spawn.stdin.writeln('Ping');
       await spawn.exitCode;
-      // TODO: https://github.com/dart-lang/sdk/issues/30119.
-      // expect(stdoutLog, ['You said: Ping', '\n']);
       expect(stdoutLog.join(''), contains('You said: Ping'));
+    });
+
+    group('should return a Process where', () {
+      test('.stdout is readable', () async {
+        final spawn = await processManager.spawn(
+          'dart',
+          arguments: [p.join('test', '_files', 'stdout_hello.dart')],
+        );
+        expect(await spawn.stdout.transform(UTF8.decoder).first, 'Hello');
+      });
+
+      test('.stderr is readable', () async {
+        final spawn = await processManager.spawn(
+          'dart',
+          arguments: [p.join('test', '_files', 'stderr_hello.dart')],
+        );
+        expect(await spawn.stderr.transform(UTF8.decoder).first, 'Hello');
+      });
     });
   });
 }